bot/requirements.txt
discord.py
requests
websockets
cryptography
